Exceptions commit

+++ UserManagmentSystem.Service/Concretes/RoleService.cs
@@ -5,6 +5,7 @@ using UserManagmentSystem.Repository.Contexts;
 using UserManagmentSystem.Service.Abstracts;
 using UserManagmentSystem.Models.Entities;
 using UserManagmentSystem.Service.Mappers.Roles;
+using UserManagmentSystem.Service.Exceptions.Types;
 
 namespace UserManagmentSystem.Service.Concretes;
 
@@ -23,6 +24,24 @@ public sealed class RoleService : IRoleService
     public void Add(RoleAddRequestDto dto)
     {
 
+        //1.Yöntem
+        bool isPresent = _roleRepository.ExistsByRoleName(dto.Name);
+        if (isPresent)
+        {
+            throw new BusinessException("Rol adı benzersiz olmalıdır.");
+        }
+
+
+        //2. Yöntem
+        //int count = _roleRepository.CountByRoleName(dto.Name);
+        //if (count > 0)
+        //{
+        //    throw new BusinessException("Rol adı benzersiz olmalıdır.");
+        //}
+
+        // İş kuralı Rol adı benzersiz olmalıdır.
+
+
         Role role = _roleMapper.CovertToEntity(dto);
 
         _roleRepository.Add(role);
@@ -31,7 +50,12 @@ public sealed class RoleService : IRoleService
     public void Delete(int id)
     {
         Role? role = _roleRepository.GetById(id);
+        if (role is null)
+        {
+            throw new NotFoundException("İlgili Rol bulunamadı.");
+        }
         //todo: Eğer ilgili rol bulunamazsa Exception Handling mekanizması
+
         _roleRepository.Delete(role!);
 
     }
@@ -49,7 +73,11 @@ public sealed class RoleService : IRoleService
     public RoleResponseDto? GetById(int id)
     {
         Role? role = _roleRepository.GetById(id);
-        //todo: Eğer ilgili rol bulunamazsa Exception Handling mekanizması
+
+        if(role is null)
+        {
+            throw new NotFoundException("İlgili Rol bulunamadı.");
+        }
         RoleResponseDto dto = _roleMapper.ConvertToResponse(role!);
 
         return dto;
